scrollable conv view w/ enter to send new msg in conv. conv list too.

+++ sync.sh
@@ -1,4 +1,8 @@
 #!/usr/bin/env bash
+if [[ $(pwd) != '/Users/dale/home/work/dalephone/phoneapp' ]]; then
+  exit 1
+fi
+
 if [[ $1 != '-f' ]]; then
   rsync -anv --delete --exclude-from=.gitignore . dale@cyberology:phone
   echo -n "Proceed? (y/n) "
